Changes based on PR comments: Rotatable x vertRotAxis, horRotAxis, lerpT, please remove abbreviations
Slidable
x Line 28, Can just be Destroy(this). the GameObject qualifier and the (Slidable) cast are unnecessary
x Line 33, it should be a serialized field rather than a GetComponent to ensure better startup time
x Lines 38 -> 43, I am unsure about those. The rigidbody has properties in the inspector, we want to completely override them? I feel like several of those properties could work with other settings, without breaking slidables. If it isn't absolutely required, we shouldn't override it's rigidbody properties.
x Right now we have a max slide speed, but its always applied 100%. I wonder if we could do a scaling speed based on distance to travel or something of that nature. If that doesn't feel good, then we should just rename to slideSpeed
x Line 123, we shouldnt set it to 0, we should set it to the velocity needed to reach the point. If you use the non-snapping one in the level the problem will be more clear. Click on one of the snappable balls, and move the mouse vey slowly, it will pop and jitter.

Both
x rotatable.cs line 26 No same line if statements.
x Properties should be TitleCase